// src/tasks/BeanstalkDeployApplication/types.gen.ts
/*!
 * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 * SPDX-License-Identifier: MIT
 */

    
/* NOTE! this file is auto-generated by generateResources.ts, do not edit it by hand. */


/*!
 * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 * SPDX-License-Identifier: MIT
 */

    
/* NOTE! this file is auto-generated by generateResources.ts, do not edit it by hand. */

export type BeanstalkDeployApplicationInput = {
    awsCredentials: string
    regionName: string
    applicationName: string
    environmentName: string
    applicationType: 'aspnet' | 'aspnetCoreWindows' | 'aspnetCoreLinux' | 's3' | 'version'
    webDeploymentArchive: string
    dotnetPublishPath: string
    deploymentBundleBucket: string
    deploymentBundleKey: string
    versionLabel: string
    description: string
    outputVariable: string
    eventPollingDelay: string
    logRequest: boolean
    logResponse: boolean
}

export type BeanstalkDeployApplicationMessages = 
    | 'DisplayApplicationType'
    | 'CreatedApplicationVersion'
    | 'CreatedApplicationVersionWithDescription'
    | 'DeployingExistingVersion'
    | 'StartingApplicationDeployment'
    | 'DeterminedBucket'
    | 'UploadingBundle'
    | 'BundleUploadCompleted'
    | 'BundleUploadFailed'
    | 'WaitingForDeployment'
    | 'EventsComing'
    | 'FailedToFindEnvironment'
    | 'FailedToDeploy'
    | 'CreatingBeanstalkBundle'
    | 'UsingExistingBeanstalkBundle'
    | 'ArchiveSize'
    | 'ZipError'
    | 'PublishingPath'
    | 'AddingAspNetCoreBundle'
    | 'AddingFilesFromDotnetPublish'
    | 'AddingManifest'
    | 'BundleComplete'
    | 'ApplicationDoesNotExist'
    | 'EnvironmentDoesNotExist'
    | 'SettingOutputVariable'
    | 'ApplicationExistsQueryError'
    | 'ApplicationExistsQueryErrorTooManyApplications'
    | 'EnvironmentExistsQueryError'
    | 'EnvironmentExistsQueryErrorTooManyEnvironments'
    | 'InvalidEventPollDelay'
    | 'ConfiguredEventPollDelay'
    | 'EventPollWaitExtended'
    | 'TaskCompleted'